Support fractional lead automation hours

+++ Dtos/LeadAutomationDtos.cs
@@ -1,4 +1,4 @@
 namespace backend.Dtos;
 
-public record LeadAutomationSettingsDto(int UnassignAfterHours, int ReminderIntervalHours);
-public record UpdateLeadAutomationSettingsRequest(int UnassignAfterHours, int ReminderIntervalHours);
+public record LeadAutomationSettingsDto(decimal UnassignAfterHours, decimal ReminderIntervalHours);
+public record UpdateLeadAutomationSettingsRequest(decimal UnassignAfterHours, decimal ReminderIntervalHours);
